fix: add @default slice to fix publisher error
- the IG publisher does not allow to use a "nested slicing" in the following form:
```
* section contains sliceA 0..*
   * section contains sliceB 0..*
```
instead, the sliceB must be contained in another actual slice, e.g.

```
* section contains sliceA 0..* and otherSlice 0..*
   * section[otherSlice] contains sliceB 0..*
```

therefore, we've added the @default slice (available from the FHIR specification) and only sliced that @default slice again, i.e.
```
* section contains sliceA 0..* and @default 0..*
   * section[@default] contains sliceB 0..*
```

unfortunately, the FHIR validator does not support @default slices as of today, so we implement a workaround using a fixed code for the discriminator (which is actually not allowed as per the FHIR specification)